split traits.rs into own folder & separate files
